Move formatters to a new file
